requirements.txt
setuptools
openai-whisper
yt-dlp
moviepy==1.0.3
google-genai
python-dotenv
tiktok-uploader
playwright
imageio==2.25.1
imageio-ffmpeg==0.4.8
